# docs/swagger.yaml
basePath: /api
definitions:
  dto.AssignCategoryRequest:
    properties:
      category_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  dto.CategoryListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.CategoryResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  dto.CategoryRequest:
    properties:
      name:
        example: Electronics
        type: string
    type: object
  dto.CategoryResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  dto.CreateOrderRequest:
    properties:
      customer_id:
        example: 123
        type: integer
      products:
        items:
          $ref: '#/definitions/dto.OrderItemRequest'
        type: array
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  dto.OrderItemRequest:
    properties:
      product_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      quantity:
        example: 2
        type: integer
      variant_id:
        description: 'Optional: order specific variant'
        example: 660e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  dto.OrderItemResponse:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
      subtotal:
        type: number
    type: object
  dto.OrderListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.OrderResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  dto.OrderResponse:
    properties:
      created_at:
        type: string
      customer_id:
        type: integer
      id:
        type: string
      payment_status:
        type: string
      products:
        items:
          $ref: '#/definitions/dto.OrderItemResponse'
        type: array
      status:
        type: string
      total_price:
        type: number
      updated_at:
        type: string
    type: object
  dto.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.UpdateOrderStatusRequest:
    properties:
      status:
        example: completed
        type: string
    type: object
...
  handler.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: marco@example.com
    name: API Support
  description: RESTful API for managing products and orders in an e-commerce system
    with payment webhook integration
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go E-Commerce API
  version: "1.0"
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: Get all categories with pagination and sorting
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - default: '"name"'
        description: Sort by field (name, created_at)
        in: query
        name: sort_by
        type: string
      - default: '"asc"'
        description: Sort order (asc, desc)
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new category (Admin only)
      parameters:
      - description: Category details
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.CategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - categories
  /orders:
    get:
      consumes:
      - application/json
      description: Get a paginated list of orders with optional filtering and sorting
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: page_size
        type: integer
      - default: '"created_at"'
        description: Sort by field (created_at, total_price)
        in: query
        name: sort_by
        type: string
      - default: '"desc"'
        description: Sort order (asc, desc)
        in: query
        name: sort_order
        type: string
      - description: Filter by status (pending, cancelled, completed)
        in: query
        name: status
        type: string
      - description: Filter by payment status (unpaid, paid, failed)
        in: query
        name: payment_status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OrderListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List all orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order with the provided products
      parameters:
      - description: Order information
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new order
      tags:
      - orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get an order by ID
      tags:
      - orders
...
  /orders/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of an existing order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update order status
      tags:
      - orders
...
  /products/{id}/categories:
    get:
      consumes:
      - application/json
      description: Get all categories assigned to a product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CategoryResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get product categories
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Assign a category to a product (Admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Category assignment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AssignCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign category to product
      tags:
      - products
  /products/{id}/categories/{category_id}:
    delete:
      consumes:
      - application/json
      description: Remove a category from a product (Admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove category from product
      tags:
      - products
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"